requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.49.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
gunicorn==20.1.0
idna==3.3
ipdb==0.13.9
ipython==8.1.1
jedi==0.18.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
oauthlib==3.2.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
sentry-sdk==1.5.7
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
Unidecode==1.3.4
urllib3==1.26.8
wcwidth==0.2.5
whitenoise==6.0.0
